Delete separate package for entity within model package
