+++ api-gateway/src/main/java/com/drbsimon/apigateway/repository/VisitorRepository.java
@@ -1,6 +1,6 @@
 package com.drbsimon.apigateway.repository;
 
-import com.drbsimon.apigateway.model.entity.Visitor;
+import com.drbsimon.apigateway.model.Visitor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
