+++ api-gateway/src/main/java/com/drbsimon/apigateway/security/service/AuthService.java
@@ -2,7 +2,7 @@ package com.drbsimon.apigateway.security.service;
 
 import com.drbsimon.apigateway.model.Gender;
 import com.drbsimon.apigateway.model.dto.UserCredentialsDTO;
-import com.drbsimon.apigateway.model.entity.Visitor;
+import com.drbsimon.apigateway.model.Visitor;
 import com.drbsimon.apigateway.service.dao.VisitorDao;
 import lombok.AllArgsConstructor;
 import lombok.Data;
